pdf_extractor/requirements.txt
# Core PDF processing
PyMuPDF>=1.23.0
pdfplumber>=0.10.0

# OCR
pytesseract>=0.3.10
Pillow>=9.0.0

# Utilities
langdetect>=1.0.9
